split out io structs and funcs
